Uploading updated learning Java

+++ tutorial/src/tutorial/Main.java
@@ -56,15 +56,17 @@ public class Main {
 		//INPUT & SCANNERS
 			// in order to get inputs you have to go to the top of the program and type import java.util.Scanner;
 		Scanner sc = new Scanner(System.in); //all this is saying is that scanner is a data type named sc. the new scanner(system.in) is typing stuff on your keyboard
-		int scanned = sc.nextInt(); //here we are creating a new variable with a data type of a string which is called scanned. and it is going to get the next stream of input from the line above
+		double scanned = sc.nextDouble(); //here we are creating a new variable with a data type of a string which is called scanned. and it is going to get the next stream of input from the line above
 								//so sc.next() returns as a string so we can't make it int scanned. if we want the user to input a int then we do sc.nextInt().
 								//so if we try to type a string here it wont work cause it can't convert the string to an int
+								//so if you want to type a double then you do sc.nextDouble() - then you have to put double 
 		boolean scanned_in = sc.nextBoolean(); 
 		
 		System.out.println(scanned);
 		// after doing this what happened was we typed hello, and then it printed hello. so basically with sc.next() we were able to input info to the keyboard.
 		// then we stored our input into variable called scanned and printing scanned
 		System.out.println(scanned_in);
+				//so i noticed that if you want to take in information then you have to comment out one of the system
 	}
 }
